fix: consolidate pkg/cli files to satisfy pkg-file-count cap
Merge group-command guard helpers into root_factory.go and fold small init and submit-batch tests into sibling test files so pkg/cli stays within the 15-file package limit after main sync.
